Bootstrap script re-worked... and working!
Everything necessairy to set up the project (except the surveys and page contents for now) on a fresh ubuntu 12.04 server install.

 * Removed _lots_ of unecessairy things from bootstrap script. Many things didn't apply to postgres databases
 * quite a bit of variables are now hardcoded in the script to save manual typing.
 * install the dependencies on the virtual environment
 * include clone of the git repo
 * install all the dependencies bot on the system with apt and on a virtual environment with pip

This can still be simplified. Data imports could be included. And a bit of interactive prompts could be eliminated, as well as some old warnings and glitches. But it is working!!!!